KR-3043 [Map Context] Map Context issues

* KR-3043 [Map Context] Map Context issues


Approved-by: Kenny Nguyen

+++ src/app/map/context-menu/cm-lock-unlock/cm-lock-unlock.service.ts
@@ -3,7 +3,7 @@ import { ToastrService } from 'ngx-toastr';
 import { HelpersService } from 'src/app/core/services/helpers/helpers.service';
 import { Store } from "@ngrx/store";
 import { Subscription } from "rxjs";
-import { selectSelectedLogicalNodes } from "../../../store/node/node.selectors";
+import { selectSelectedLogicalNodes, selectSelectedPhysicalNodes } from "../../../store/node/node.selectors";
 import { selectSelectedPortGroups } from "../../../store/portgroup/portgroup.selectors";
 import { selectSelectedMapImages } from "../../../store/map-image/map-image.selectors";
 import { selectSelectedMapLinks } from "../../../store/map-link/map-link.selectors";
@@ -19,6 +19,8 @@ export class CMLockUnlockService implements OnDestroy {
   selectSelectedMapLinks$ = new Subscription();
 
   selectedNodes: any[] = [];
+  selectedLogicalNodes: any[] = [];
+  selectedPhysicalNodes: any[] = [];
   selectedPortGroups: any[] = [];
   selectedMapImages: any[] = [];
   selectedMapLinks: any[] = [];
@@ -30,7 +32,12 @@ export class CMLockUnlockService implements OnDestroy {
   ) {
     this.selectSelectedLogicalNodes$ = this.store.select(selectSelectedLogicalNodes).subscribe(nodes => {
       if (nodes) {
-        this.selectedNodes = nodes;
+        this.selectedLogicalNodes = nodes;
+      }
+    });
+    this.selectSelectedLogicalNodes$ = this.store.select(selectSelectedPhysicalNodes).subscribe(nodes => {
+      if (nodes) {
+        this.selectedPhysicalNodes = nodes;
       }
     });
     this.selectSelectedPortGroups$ = this.store.select(selectSelectedPortGroups).subscribe(portGroups => {
@@ -57,53 +64,63 @@ export class CMLockUnlockService implements OnDestroy {
     this.selectSelectedMapLinks$.unsubscribe();
   }
 
-  getLockMenu(cy: any) {
+  getLockMenu(cy: any, mapCategory: any) {
     return {
       id: "lock_node",
       content: "Lock",
       selector: "node",
       onClickFunction: (event: any) => {
-        this.lockNodes(cy);
-        this.toastr.success("Locked the nodes");
+        this.lockNodes(cy, mapCategory);
       },
       hasTrailingDivider: false,
       disabled: false,
     }
   }
 
-  getUnlockMenu(cy: any) {
+  getUnlockMenu(cy: any, mapCategory: any) {
     return {
       id: "unlock_node",
       content: "Unlock",
       selector: "node",
       onClickFunction: (event: any) => {
-        this.unlockNodes(cy);
-        this.toastr.success("Unlocked the nodes");
+        this.unlockNodes(cy, mapCategory);
       },
       hasTrailingDivider: true,
       disabled: false,
     }
   }
 
-  lockNodes(cy: any) {
+  lockNodes(cy: any, mapCategory: any) {
+    this.selectedNodes = mapCategory === 'logical' ? this.selectedLogicalNodes : this.selectedPhysicalNodes;
     this.selectedNodes.concat(this.selectedPortGroups, this.selectedMapImages, this.selectedMapLinks).forEach((ele: any) => {
       const selectedEle = cy.getElementById(ele.data.id);
-      selectedEle.data('locked', true);
-      selectedEle.lock();
-      const d = selectedEle.data();
-      d.updated = true;
-      this.helpersService.addBadge(cy, selectedEle);
+      if (selectedEle.locked()) {
+        this.toastr.warning('Already locked');
+      } else {
+        selectedEle.data('locked', true);
+        selectedEle.lock();
+        const d = selectedEle.data();
+        d.updated = true;
+        this.helpersService.addBadge(cy, selectedEle);
+        this.toastr.success("Locked the nodes");
+      }
     });
   }
 
-  unlockNodes(cy: any) {
+  unlockNodes(cy: any, mapCategory: any) {
+    this.selectedNodes = mapCategory === 'logical' ? this.selectedLogicalNodes : this.selectedPhysicalNodes;
     this.selectedNodes.concat(this.selectedPortGroups, this.selectedMapImages, this.selectedMapLinks).forEach((ele: any) => {
       const selectedEle = cy.getElementById(ele.data.id);
-      selectedEle.data('locked', false);
-      selectedEle.unlock();
-      const d = selectedEle.data();
-      d.updated = true;
-      this.helpersService.removeBadge(selectedEle);
+      if (!selectedEle.locked()) {
+        this.toastr.warning('Already Unlocked');
+      } else {
+        selectedEle.data('locked', false);
+        selectedEle.unlock();
+        const d = selectedEle.data();
+        d.updated = true;
+        this.helpersService.removeBadge(selectedEle);
+        this.toastr.success("Unlocked the nodes");
+      }
     });
   }
 }
